Remove asserter interface
